Add function control flow graph API
Added the API for the function CFG to the backend. Note, that
files are now NOT deleted, as it has to be around for function
CFG analysis to happen upon the users requests. Pre-cfg-analyzing
every function in the file is extremely costly, and for a 1mb file
will take several minutes. No good!

The func_cfg function now populates, for every block in the specified
function, the src and dst many-to-many fields, effectively creating
an directed graph within the block models. This will then need to be
rebuilt/parsed on the frontend in order to display the CFG. That's
an issue for tomorrow.

PS. Lots of code quality issues. Main priority is to get the CFG
working on the frontend with React Flow. Once that is done, extra
debugging statements need to be removed. API needs to be standardized.
Functions and variables need to be renamed consistently. Code needs
to have redundant portions removed/made prettier. I've made a good
amount of work for future me :)