Function CFG returns node and edges object of function
Task needs to be reworked to support returning information for
multiple asynchronous requests. As of now it only handles file upload
and returning a project ID, however we need it for returning CFG info
as well, and potentially other things in the future, such as disasm
on request (if one-shot analyzing all functions is too expensive).