Add automatic file deletion
Add celery beat periodic task. In the future, this will be used
for periodic file eviction, if needed. Auto file deletion based on
total size may make this not needed, depending on the file turnover,
i.e. if enough files are uploaded within a 2 day period to not
warrant deleting periodically. This remains to be seen. I'll
implement this in the future, and will be used if needed.